refactor: 彻底删除废弃的executionContext.js - 完成架构统一
🗑️ 清理废弃代码：
- 删除 executionContext.js 文件（221行代码）
- 重构 MCPServerStdioCommand.js 使用新架构
- 统一使用 DirectoryService + ServerEnvironment + ProjectManager

🎯 新架构实现：
- getExecutionContext(): 简化的同步实现，移除复杂查找逻辑
- initializeDebugInfo(): 异步调试信息，使用DirectoryService
- 保持完整的向后兼容性和功能覆盖

⚡ 性能优化：
- 移除216行废弃代码和复杂的目录遍历逻辑
- 统一使用异步API，避免同步阻塞
- 更好的错误处理和跨平台支持

✅ 验证完成：
- MCP服务器启动正常
- 所有功能测试通过
- DirectoryService调试信息正常
- 项目路径管理完全正常

🏗️ 架构成果：
- 100% 统一新架构，零废弃代码
- 奥卡姆剃刀原则：最简架构实现最强功能
- 完全符合"一次设置，全程使用"的设计理念
